Challenge 2 Folder/requirements.txt
fastapi
motor
pydantic
uvicorn
pytesseract
Pillow
transformers